Move views into package
Refs #34
