+++ resources/tests/test_resource_image_view.py
@@ -4,7 +4,7 @@ from django.core.urlresolvers import reverse
 from django.utils.six import BytesIO
 from PIL import Image
 
-from resources.images import parse_dimension_string
+from resources.views.images import parse_dimension_string
 from resources.tests.utils import create_resource_image
